Enable RGBA32 for texture data and removing an 8 bit limit on the att…

+++ src/renderers/webgl/WebGLState.js
@@ -15,7 +15,7 @@ THREE.WebGLState = function ( gl, extensions, paramThreeToGL ) {
 	var maxVertexAttributes = gl.getParameter( gl.MAX_VERTEX_ATTRIBS );
 	var newAttributes = new Uint8Array( maxVertexAttributes );
 	var enabledAttributes = new Uint8Array( maxVertexAttributes );
-	var attributeDivisors = new Uint8Array( maxVertexAttributes );
+	var attributeDivisors = new Uint32Array( maxVertexAttributes );
 
 	var capabilities = {};
